refactor: update imports to use centralized schemas
* Changed imports for `ApiError`, `FileEntryListSchema`, and `UCDStoreManifestSchema` to use the new centralized schema location.
* Removed unused `schemas.ts` file and cleaned up import paths in various files.
* Improved organization and maintainability of error handling and schema definitions across the project.
