--- apps/api/src/lib/index.ts
@@ -1,19 +1,0 @@
-export { clearCacheEntry } from "./cache";
-export {
-  badGateway,
-  badRequest,
-  customError,
-  forbidden,
-  internalServerError,
-  notFound,
-} from "./errors";
-export type { CustomResponseOptions, ResponseOptions } from "./errors";
-
-export { errorHandler, notFoundHandler } from "./handlers";
-export { createResponseComponentBuilder } from "./openapi";
-
-export type { ApiError } from "./schemas";
-
-export { ApiErrorSchema } from "./schemas";
-export { setupCors, setupRatelimit } from "./setups";
-export { strictJSONResponse } from "./strict";
