--- apps/api/src/routes/v1_files/schemas.ts
@@ -1,8 +1,0 @@
-import { z } from "@hono/zod-openapi";
-import {
-  UCDStoreManifestSchema as _UCDStoreManifestSchema,
-  FileEntrySchema,
-} from "@ucdjs/schemas";
-
-export const UCDStoreManifestSchema = _UCDStoreManifestSchema.openapi("UCDStoreManifest");
-export const FileEntryListSchema = z.array(FileEntrySchema).openapi("FileEntryList");
