+++ apps/api/src/worker.ts
@@ -2,7 +2,8 @@ import type { HonoEnv } from "./types";
 import { OpenAPIHono } from "@hono/zod-openapi";
 import { Scalar } from "@scalar/hono-api-reference";
 import { env } from "hono/adapter";
-import { errorHandler, notFoundHandler, setupCors, setupRatelimit } from "./lib";
+import { errorHandler, notFoundHandler } from "./lib/handlers";
+import { setupCors, setupRatelimit } from "./lib/setups";
 import { buildOpenApiConfig, registerApp } from "./openapi";
 import { V1_FILES_ROUTER } from "./routes/v1_files/routes";
 import { V1_VERSIONS_ROUTER } from "./routes/v1_versions/routes";
