+++ apps/api/test/unit/errors.test.ts
@@ -1,4 +1,4 @@
-import type { ApiError } from "../../src/lib";
+import type { ApiError } from "@ucdjs/schemas";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import {
   badGateway,
@@ -6,7 +6,7 @@ import {
   customError,
   internalServerError,
   notFound,
-} from "../../src/lib";
+} from "../../src/lib/errors";
 
 const mockDate = new Date("2023-06-15T10:30:00.000Z");
