+++ packages/fetch/src/components.ts
@@ -6,7 +6,6 @@
 
 import type { components } from "./.generated/api";
 
-export type ApiError = components["schemas"]["ApiError"];
 export type UnicodeVersionList = components["schemas"]["UnicodeVersionList"];
 export type UnicodeTreeNode = components["schemas"]["UnicodeTreeNode"];
 export type UnicodeTree = components["schemas"]["UnicodeTree"];
