+++ packages/schemas/src/index.ts
@@ -1,5 +1,8 @@
-export { FileEntrySchema, UCDStoreManifestSchema } from "./fs";
-export type { FileEntry, UCDStoreManifest } from "./fs";
+export { ApiErrorSchema } from "./api";
+export type { ApiError } from "./api";
 
-export { UnicodeVersionSchema } from "./unicode";
-export type { UnicodeVersion } from "./unicode";
+export { FileEntryListSchema, FileEntrySchema, UCDStoreManifestSchema } from "./fs";
+export type { FileEntry, FileEntryList, UCDStoreManifest } from "./fs";
+
+export { UnicodeVersionListSchema, UnicodeVersionSchema } from "./unicode";
+export type { UnicodeVersion, UnicodeVersionList } from "./unicode";
